moved VariableInstruction into instruction package
