+++ src/main/java/org/betonquest/betonquest/id/VariableID.java
@@ -1,10 +1,10 @@
 package org.betonquest.betonquest.id;
 
-import org.betonquest.betonquest.VariableInstruction;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
+import org.betonquest.betonquest.instruction.VariableInstruction;
 import org.jetbrains.annotations.Nullable;
 
 /**
